requirements.txt
feedparser>=6.0.0
jinja2>=3.1.0
requests>=2.31.0
schedule>=1.2.0
python-dateutil>=2.8.0
